requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
